+++ gitnexus/src/cli/analyze-config.ts
@@ -30,7 +30,7 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import type { AnalyzeOptions } from './analyze.js';
+import type { AnalyzeOptions } from './analyze-options.js';
 
 export const GITNEXUS_RC_FILENAME = '.gitnexusrc';
