+++ gitnexus/src/core/ingestion/di-extractors/index.ts
@@ -15,56 +15,14 @@
  */
 
 import { SupportedLanguages } from 'gitnexus-shared';
-import type { GraphNode } from 'gitnexus-shared';
+import type { DiResolver } from './types.js';
 import { springDiResolver } from './spring.js';
 
-/** A successful injection-site match, produced by a per-language resolver. */
-export interface DiInjectionMatch {
-  /** The requested dependency type name. */
-  targetTypeName: string;
-  /** A collection receives every matching provider; a single site may need
-   *  framework-specific named/preferred-provider disambiguation. */
-  cardinality: 'single' | 'collection';
-  /** Statically known provider name requested at the injection site. The
-   *  resolver owns the human-readable explanation of that selection. */
-  namedSelection?: {
-    name: string;
-    reason: string;
-    /** Name-first frameworks may fall back to type only for implicit/default
-     * names. Explicit names remain strict. */
-    fallbackToType?: boolean;
-  };
-  /** Most injection edges originate at the owning Class. Factory-method
-   * parameters preserve the Method as the semantic source. */
-  edgeSource?: 'owner-class' | 'site';
-  /** Human-readable edge reason. Framework specifics (names, idioms,
-   *  collection wrapper, gating annotation) live in this payload so the
-   *  shared `di` phase stays framework-neutral. */
-  reason: string;
-}
-
-/** Provider metadata used by the shared resolver without naming a framework. */
-export interface DiProviderMatch {
-  /** Provider names and aliases that can satisfy a named injection. */
-  names: readonly string[];
-  /** Optional type directly provided by a declaration node, such as a
-   * framework factory method whose node is not itself a Class. */
-  providedTypeName?: string;
-  /** Graph node that declares this provider. The shared phase excludes a
-   * provider from injection into its own declaration site without knowing the
-   * framework-specific declaration model. */
-  declaredByNodeId?: string;
-  /** Present when the framework marks this as its preferred candidate. The
-   *  value is appended to the emitted edge reason when it disambiguates. */
-  preferenceReason?: string;
-}
-
-/** Per-language DI behavior. Matchers receive whole nodes so the shared phase
- * remains ignorant of language/framework-specific property shapes. */
-export interface DiResolver {
-  matchInjectionSites(node: GraphNode): readonly DiInjectionMatch[];
-  matchProvider(node: GraphNode): DiProviderMatch | null;
-}
+/** The resolver contract lives in the leaf `./types.js` so an implementation
+ *  can depend on it without depending on this registry (which imports every
+ *  implementation). Re-exported here so consumers can keep importing the
+ *  contract and the registry from one place. */
+export type { DiInjectionMatch, DiProviderMatch, DiResolver } from './types.js';
 
 /** All `SupportedLanguages` string values, for narrowing raw graph strings. */
 const SUPPORTED_LANGUAGE_VALUES: ReadonlySet<string> = new Set(Object.values(SupportedLanguages));
