+++ gitnexus/src/core/ingestion/scope-resolution/workspace-index.ts
@@ -40,30 +40,14 @@
  */
 
 import type { ParsedFile, Scope, ScopeId, ScopeTree, SymbolDefinition } from 'gitnexus-shared';
+import type { WorkspaceResolutionIndex } from './workspace-index-types.js';
 import { isClassLike } from './scope/walkers.js';
 
-export interface WorkspaceResolutionIndex {
-  /** Class def `nodeId` → that class's `Scope`. */
-  readonly classScopeByDefId: ReadonlyMap<string, Scope>;
-
-  /** Inverse of `classScopeByDefId`: class `Scope.id` → class def `nodeId`.
-   *  Built in the same pass; used by the implicit-`this` overload picker
-   *  in `free-call-fallback.ts` to skip an O(C) reverse scan. */
-  readonly classScopeIdToDefId: ReadonlyMap<ScopeId, string>;
-
-  /** Module scope by file path. */
-  readonly moduleScopeByFile: ReadonlyMap<string, Scope>;
-
-  /** Precomputed `simpleName → first module-local callable def` (the
-   *  workspace-wide fallback of `findExportedDefByName`). Materialized here
-   *  ONCE from the resident module scopes so that fallback is an O(1) lookup
-   *  instead of an O(files) scan over every module scope's bindings on each
-   *  unresolved free call — which, under the disk-backed scopeTree, would
-   *  otherwise fault every module scope in from disk per call (the throughput
-   *  killer). "First module-local callable in `moduleScopeByFile` order" is the
-   *  exact semantics the old scan returned, so it is byte-identical. */
-  readonly exportedCallableByName: ReadonlyMap<string, SymbolDefinition>;
-}
+/** The index *shape* lives in the leaf `./workspace-index-types.js` so
+ *  `scope/walkers.ts` — which this builder calls into — can type against it
+ *  without importing this module back. Re-exported here so consumers keep
+ *  importing the type and the builder from one place. */
+export type { WorkspaceResolutionIndex } from './workspace-index-types.js';
 
 /**
  * A `ReadonlyMap<K, Scope>` view backed by a `K → ScopeId` map plus a
