+++ gitnexus/src/storage/branch-index.ts
@@ -2,16 +2,21 @@
  * Branch-index primitives (#2106).
  *
  * Extracted from `repo-manager.ts` to keep the multi-branch slug/placement
- * logic in one focused module. `getStoragePaths`, `loadMeta`, and the registry
- * I/O stay in `repo-manager.ts`; this module imports the two it needs at
- * call-time only (no module-load cross-calls), so the repo-manager ⇄
- * branch-index import cycle is ESM-safe. `repo-manager.ts` re-exports these so
- * existing import sites keep working unchanged.
+ * logic in one focused module. The registry I/O and the metadata WRITE side
+ * stay in `repo-manager.ts`, which re-exports these so existing import sites
+ * keep working unchanged.
+ *
+ * The metadata READ primitives this module needs (`getStoragePath`, `loadMeta`,
+ * `RepoMeta`) come from `repo-meta.ts`, a leaf below both modules — NOT from
+ * `repo-manager.ts`. Importing them from there made the two modules import
+ * values out of each other, and the only thing keeping that ESM-safe was that
+ * neither side called across at module-evaluation time. Reading from the layer
+ * below removes the cycle instead of depending on that timing.
  */
 
 import { createHash } from 'crypto';
 import { sanitizeRepoName } from './git.js';
-import { getStoragePaths, loadMeta, type RepoMeta } from './repo-manager.js';
+import { getStoragePath, loadMeta, type RepoMeta } from './repo-meta.js';
 
 /**
  * Per-branch index summary nested under a registry entry (#2106). Records
@@ -66,7 +71,10 @@ export const resolveBranchPlacement = async (
 ): Promise<{ branch?: string }> => {
   // Detached HEAD / non-git / no label → flat (CI-safe, byte-identical).
   if (!label) return {};
-  const { storagePath } = getStoragePaths(repoPath);
+  // The flat slot only — identical to `getStoragePaths(repoPath).storagePath`,
+  // which is `getStoragePath(repoPath)` verbatim (the `branch` argument only
+  // ever scopes `lbugPath`/`metaPath`, never `storagePath`).
+  const storagePath = getStoragePath(repoPath);
   const flatMeta = await loadMeta(storagePath);
   // The flat slot's owner is authoritative ONLY when it is a non-empty string.
   // A corrupt/hand-edited meta (empty string, or a non-string value that slips
